requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
pytest==8.2.2
spotipy==2.24.0
torch==2.3.1
transformers==4.42.3